Implement commenter profile image fetching
Fetch the profile pictures of commenters in the CommentsFragment.

(Mulled over using some sort of WeakHashMap similar to LazyList, but feel
it may be overkill for this scenario.  Reading suggests it's not an
effective caching mechanism, so the simplicity of the current approach
outweighs any memory benefits of WeakReferences for the time being.)