Fix notification service logic
This was updating the most recent photo id every time it found new
photos.

Hence if it first found 3 new photos, and alerted for that, and then the
next time round found 1 more, it would overwrite the notification for
the 3 with a notification for just the latest one.

Part of a fix is that there is also no need to have the Service keep or
update a most recent id, it can simply check the new list against the
last viewed id.  After this it just needed to record the most recent
*notified* id in order to avoid duplicate notifications.  Additional
photos can still be found and the notification updated.

Also:
* Cancel alarms in MainActivity if the preference is disabled, in case
  they are still set when entering.

* Tidy up Constants a little.