Move menu from PhotoViewerActivity into Fragment
Based on the previous pattern of moving the menu code into the
fragments, also move PhotoViewerActivity's menu management code into
PhotoViewerFragment.

This *greatly* simplifies management of the favorite icon as we don't
have to try and track which fragment is showing (which was difficult and
buggy).  Each fragment just manages it's own favorite button.