Fix user management across activities
The current User object was not being restored properly if an Activity
was destroyed.

This is a tricky one to solve.  The fragments could manage the User
object, but they have no knowledge whether they are profile fragments or
owner fragments, so you end up with the wrong User object being
restored in the wrong fragment type.

It seems best then to have the host Activity manage it, though this
results in some extra coupling between the fragment and host activity.
I tried to alleviate this through using an interface, but I ran into
NPEs from the fragments losing their reference to the interface
onResume.  Seems to me Android lifecycle management is annoying to
manage in general.

Also:
* Fix members in fragments that were not been stored/restored
properly after fragment teardown.  Bugs around this were discovered
using the "Don't keep activities" option under developer options.