Workaround FragmentTransaction bug by moving menu
There's a bug where fragment transactions for the comments/exif throw
exceptions after a rotation.  Calling commit throws an IllegalStateException
which is expected, but commitAllowingStateLoss throws an exception about the
activity having been destroyed, even though I update the fragment's reference
to it in BaseFragment's onResume.

So, instead of having the fragment call up to the activity to do the hide/show
transaction, I'm just giving the PhotoViewerActivity it's own menu.

The original workaround for this was to override onSaveInstanceState in
PhotoViewerActivity to not call up to super(), but then you lose all the auto
persist help.

Conflicts:

	res/menu/photoviewer_menu.xml
	src/com/bourke/glimmr/activities/PhotoViewerActivity.java
	src/com/bourke/glimmr/fragments/viewer/PhotoViewerFragment.java