Swap context menu for longclicklistener+dialog
Was suffering from a problem where each fragment in the viewpager was
registering for a context menu, potentially overwriting the current one.
This left a situation where the context menu shown was referring to the
wrong fragment.

So I have created a DialogFragment to mimic the context menu and show it
with a longclicklistener on the grid items.

Slightly more code but seems to work well.