Move oauth callback into settings
Fix a nasty bug caused by user selecting wrong app to complete
authorisation.

Using a different callback url for free/pro versions avoids the intent
dialog and automatically selects the right version of the app to
complete the login.

Moved it to a string resource so it could be referenced from the
AndroidManifest avoiding having to update it in two places.