Refactor AppService to support multiple handlers
Define an interface for a handler that can be run periodically via
AppService.  This allows for just one Service that can delagate it's
work.

Also makes way for showing flickr activity items in the menu drawer.
They are currently fetched via an AppService handler, and stored to
disk for MainActivity to read.

Next steps should be:
* Define a new row type to display activity items in the MenuDrawer.

* Provide a way to refresh these manually, as well as the background
  service refreshing them periodically.

Also:
* Add commit id to CWACWakeful jar

* Update flickrj-android jar with bug fix for Activity and make activity
  classes serializable

Conflicts:
	src/com/bourke/glimmr/activities/MainActivity.java
	src/com/bourke/glimmr/services/AppService.java