Ensure menu drawer stays fresh with cache
Each time ActivityNotificationHandler.storeItemList updates the item
list json, it also stores a timestamp to SharedPreferences.

When the menu drawer is initalised, it writes it's own time stamp.  Each
time the drawer is opened subsequently, it compares if the
ActivityNotificationHandler's stamp is newer than it's own.  If so it
reloads it's items from the json.