Improve ExploreActivity management on login
Use startActivityForResult/setResult to allow ExploreActivity to pass a
result back to MainActivity if it needs to exit.  This prevents a blank
MainActivity showing and the user having to press back twice to exit the
app when not logged in.

Also add intent flags when starting the MainActivity after logging in,
to help prevent the user navigating back to the login screen once
authorised.