Remove sets functionality from uploader
This one is a more complicated then it should be, as Flickr don't let
you add this info in the upload api call.

Hence to do this you would have to upload the photos, wait till they're
added, then fetch the ids and kick off new tasks to update the sets.

Also:
* Refactor the tape queue files into constants so it's clear every part
  of the code will use the same queues.