Show dialog instead of crash when null oauth
Play console report: http://goo.gl/QZHic0

Common crash on play store is that the GetRequestToken task returns
null.  This is due to some exception, possibly network related or from
Flickr.  There's not much we can do here other than show a dialog asking
to try again, or contact me in the case it persists, there may be
something about that users account that doesn't fit in with the expected
auth flow.