Run DownloadPhotoTask on tape service
Common crash on Play Store is an IllegalStateException from
DownloadPhotoTask calling it's callback when the user has moved on and
closed the PhotoViewerActivity.

Cancelling this task when the user navigates away is not a good idea as
they're assuming success.  Hence, we should run it in a service, and use
the already established tape framework to accomplish retries.