Begin to abstract photolist data into model
Begin to abstract the logic that manages the tasks and list of photos
(or other items) into a data model component.

When finished this should solve a common user complaint, in that the
PhotoViewerActivity can't fetch the next page of photos when they reach
the end.

With the management of this in a seperate component, both the
PhotoGridFragments and PhotoViewerActivity can request the list to be
updated, and will avoid duplicating code.

The change unfortunately breaks a lot of current contracts so a fair bit
of refactoring will be needed.

http://programmers.stackexchange.com/q/232691/123678